Keep client alive when connection to remote server is lost.
The usecase is the new message handler backtrace dialog. Before
the dialog about connection loss must be ignored by the user, as
closing it would kill the whole client. This way he can accept it,
then take a look at the backtrace dialog.

The central widget is disabled though as interaction with it will
crash sooner or later since most assertions don't hold anymore if
we are not connected to a remote server. This will need more work.